% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymode.R
\name{mode_numeric}
\alias{mode_numeric}
\title{Calculate the Mode of a Numeric Vector}
\usage{
mode_numeric(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value representing the mode of \code{x}.
}
\description{
This function computes the statistical mode of a numeric vector.
It returns the value that occurs most frequently in the input.
If multiple values share the highest frequency, the first such
value (in sorted order) is returned.
}
\examples{
mode_numeric(c(1, 2, 2, 3))

}
