% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_density.R
\name{myhdi}
\alias{myhdi}
\title{myhdi function}
\usage{
myhdi(funame, credwidth = 0.95, tol = 1e-08, ...)
}
\arguments{
\item{funame}{a function call}

\item{credwidth}{a decimal value}

\item{tol}{a numeric input}

\item{...}{other optional values}
}
\value{
list of values
}
\description{
myhdi function
}
\examples{
myhdi(funame = qbeta , shape1 = 5, shape2 = 7)
}
