% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysd.R
\name{mysd}
\alias{mysd}
\title{My Standard Deviation calculator}
\usage{
mysd(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{logical, remove NA values from vector}
}
\value{
standard deviation as float
}
\description{
My Standard Deviation calculator
}
\examples{
mysd(c(1,2,3,4,5))
}
